requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.7
cffi==1.14.0
Django==3.0.5
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-markdownx==3.0.1
olefile==0.44
Pillow==7.1.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2017.2
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.0.1
